unused imports removed and elements moved
