+++ ui/src/components/RoutesMap.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { Routes, Route } from 'react-router-dom';
 import routes from '../routes';
