+++ ui/src/components/elements/AuctionListItem/AuctionListItem.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import ListItem from '@material-ui/core/ListItem';
 import ListItemText from '@material-ui/core/ListItemText';
