+++ ui/src/components/enums/AuctionType.ts
