+++ ui/src/pages/EditAuctionPage/EditAuctionPage.tsx
@@ -14,9 +14,9 @@ import { useMutation } from 'react-query';
 import { toast } from 'react-toastify';
 import { useNavigate } from 'react-router-dom';
 import getAuction from './services/getAuction';
-import { AuctionType } from '../CreateAuctionPage/enums/AuctionType';
+import { AuctionType } from '../../components/enums/AuctionType';
 import editAuction from './services/editAuction';
-import SearchInput from '../Home/components/Search/Search';
+import SearchInput from '../../components/elements/Search/Search';
 import ItemList from './components/ItemList/ItemList';
 import deleteAuctionItem from './services/deleteAuctionItem';
 import AddIcon from '@mui/icons-material/Add';
