+++ ui/src/pages/MyAuctionsPage/MyAuctionsPage.tsx
@@ -1,8 +1,8 @@
 import axios from 'axios';
 import { useState } from 'react';
 import { useQuery } from 'react-query';
-import AuctionList from '../Home/components/AuctionList/AuctionList';
-import SearchInput from '../Home/components/Search/Search';
+import AuctionList from '../../components/elements/AuctionList/AuctionList';
+import SearchInput from '../../components/elements/Search/Search';
 import { ToastContainer } from 'react-toastify';
 import useAuctionNextCursor from '../../hooks/useAuctionNextCursor';
 import getTokenFromCookies from '../../components/utils/getTokenFromCookies';
