+++ ui/src/pages/Welcome/Welcome.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { Link } from 'react-router-dom';
 
 export default function Welcome() {
